Add parameter support for native functions.
This allows passing parameters to native functions, for now we pass
them as std::vector<Value> which feels naive, but until I come up with
a better way to do it, this is how it works.

We also now have a print() function that takes an arbitrary number
of strings and prints them, this will help tremendously while debugging.