Add support for FFI :^)
We can now use foreign functions inside the ore interpreter!
This introduces the import function, for now it allows importing linux
shared objects and returns an FFIObject.

When initializing an FFIObject, we look for a function called
OreInitialize() and pass it a reference to an exports entries vector
to populate, this way we know which functions are available and under
what name, any functions not explicitly exported are simply ignored.