Implement and migrate to Ore::Result
This introduces a new Ore::Result type that will serve as the return
type for ASTNode::execute(), it encapsulates all the types you can
return from executing a statement (normal value, returning a value,
breaking from a loop, continuing in a loop, or throwing an exception).

This allows migrating from keeping track of the exception object inside
the interpreter and allows easily passing it through return values, also
reduces the clutter related to unwinding blocks in the interpreter.

To handle exception flow through-out the AST execution, we now have a
TRY() macro that utilizes the "statements in expressions" GCC and clang
extension; now you can TRY(node.execute(interpreter)) and it will
automatically return from the function if the result is an exception.