Make the garbage collector conservative
If we happen to call collect_garbage() while in the middle of
constructing a new object, it will get collected since we don't have a
reference to it yet.
This change prevents that by scanning the values of the registers and
the program stack to find references to potential objects under
construction and marking them to prevent their collection.