Check property existance in object with using Object::contains()
Previously, to get a property from an object, we would iterate over the
object shapes twice, once for Object::contains() and once for
Object::get().
This optimization changes that behavior into checking it only twice in
Object::get() and throwing a ReferenceException if it does not exist,
which will reduce the runtime of getting a variable significantly.